# docs/seo-brain/openapi.v1.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "SEO Brain API",
  "version": "0.2.0"
 },
 "paths": {
...
  "/api/v1/sites/{site_id}": {
   "get": {
    "tags": [
     "sites"
    ],
    "summary": "Get Site",
    "operationId": "get_site_api_v1_sites__site_id__get",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Site Api V1 Sites  Site Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "sites"
    ],
    "summary": "Update Site",
    "operationId": "update_site_api_v1_sites__site_id__patch",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SiteUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Update Site Api V1 Sites  Site Id  Patch"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "sites"
    ],
    "summary": "Delete Site",
    "description": "Delete a site. Refuses (409 `site_has_data`) when related rows exist unless force=true.\nFiles in the workspace directory are never deleted by the API.",
    "operationId": "delete_site_api_v1_sites__site_id__delete",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "force",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "description": "also delete all site data (graph, GSC, crawl, memory)",
       "default": false,
       "title": "Force"
      },
      "description": "also delete all site data (graph, GSC, crawl, memory)"
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Delete Site Api V1 Sites  Site Id  Delete"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/sites/{site_id}/generation/meta": {
   "get": {
    "tags": [
     "generation"
    ],
    "summary": "Meta",
    "operationId": "meta_api_v1_sites__site_id__generation_meta_get",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Meta Api V1 Sites  Site Id  Generation Meta Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/sites/{site_id}/generation/memory-preview": {
   "get": {
    "tags": [
     "generation"
    ],
    "summary": "Memory Preview",
    "operationId": "memory_preview_api_v1_sites__site_id__generation_memory_preview_get",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Memory Preview Api V1 Sites  Site Id  Generation Memory Preview Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/sites/{site_id}/content/{cid}/generate/estimate": {
   "post": {
    "tags": [
     "generation"
    ],
    "summary": "Estimate",
    "operationId": "estimate_api_v1_sites__site_id__content__cid__generate_estimate_post",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "cid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cid"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/GenerateBody"
         },
         {
          "type": "null"
         }
        ],
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Estimate Api V1 Sites  Site Id  Content  Cid  Generate Estimate Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/sites/{site_id}/content/{cid}/generate": {
   "post": {
    "tags": [
     "generation"
    ],
    "summary": "Generate",
    "description": "Start a section-by-section generation run as a job. Returns the run (poll GET /generation/runs/{run_id} or stream).",
    "operationId": "generate_api_v1_sites__site_id__content__cid__generate_post",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "cid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cid"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/GenerateBody"
         },
         {
          "type": "null"
         }
        ],
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Generate Api V1 Sites  Site Id  Content  Cid  Generate Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/sites/{site_id}/generation/runs": {
   "get": {
    "tags": [
     "generation"
    ],
    "summary": "Runs",
    "operationId": "runs_api_v1_sites__site_id__generation_runs_get",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "content_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Content Id"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": true
         },
         "title": "Response Runs Api V1 Sites  Site Id  Generation Runs Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/sites/{site_id}/generation/runs/{run_id}": {
   "get": {
    "tags": [
     "generation"
    ],
    "summary": "Run",
    "operationId": "run_api_v1_sites__site_id__generation_runs__run_id__get",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Run Id"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Run Api V1 Sites  Site Id  Generation Runs  Run Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/sites/{site_id}/generation/runs/{run_id}/stream": {
   "get": {
    "tags": [
     "generation"
    ],
    "summary": "Stream",
    "description": "SSE progress: step_start / step_done / plan / done / failed / cancelled (+ keepalive). Backlog replayed for late subscribers.",
    "operationId": "stream_api_v1_sites__site_id__generation_runs__run_id__stream_get",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Run Id"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/sites/{site_id}/generation/runs/{run_id}/accept": {
   "post": {
    "tags": [
     "generation"
    ],
    "summary": "Accept",
    "description": "Manual mode: human promotes the assembled article to a draft version (+ score/review). Never automatic.",
    "operationId": "accept_api_v1_sites__site_id__generation_runs__run_id__accept_post",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Run Id"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Accept Api V1 Sites  Site Id  Generation Runs  Run Id  Accept Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/sites/{site_id}/generation/runs/{run_id}/cancel": {
   "post": {
    "tags": [
     "generation"
    ],
    "summary": "Cancel",
    "operationId": "cancel_api_v1_sites__site_id__generation_runs__run_id__cancel_post",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Run Id"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Cancel Api V1 Sites  Site Id  Generation Runs  Run Id  Cancel Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/sites/{site_id}/content/{cid}/agents/{agent}/run": {
   "post": {
    "tags": [
     "generation"
    ],
    "summary": "Run Agent",
    "description": "Run a single agent (research/outline/seo/linking/reviewer) as a proposal — no draft is created.",
    "operationId": "run_agent_api_v1_sites__site_id__content__cid__agents__agent__run_post",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "cid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cid"
      }
     },
     {
      "name": "agent",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Agent"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/GenerateBody"
         },
         {
          "type": "null"
         }
        ],
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Run Agent Api V1 Sites  Site Id  Content  Cid  Agents  Agent  Run Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/sites/{site_id}/content/{cid}/feedback": {
   "post": {
    "tags": [
     "generation"
    ],
    "summary": "Feedback",
    "description": "Human rating 1–5 + tags on a draft/run (learning signal; never changes routing).",
    "operationId": "feedback_api_v1_sites__site_id__content__cid__feedback_post",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "cid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cid"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/FeedbackBody"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Feedback Api V1 Sites  Site Id  Content  Cid  Feedback Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "generation"
    ],
    "summary": "List Feedback",
    "operationId": "list_feedback_api_v1_sites__site_id__content__cid__feedback_get",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "cid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cid"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": true
         },
         "title": "Response List Feedback Api V1 Sites  Site Id  Content  Cid  Feedback Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
...
   "FeedbackBody": {
    "properties": {
     "rating": {
      "type": "integer",
      "maximum": 5.0,
      "minimum": 1.0,
      "title": "Rating"
     },
     "tags": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Tags",
      "default": []
     },
     "draft_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Draft Id"
     },
     "run_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Run Id"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     }
    },
    "type": "object",
    "required": [
     "rating"
    ],
    "title": "FeedbackBody"
   },
   "GenerateBody": {
    "properties": {
     "mode": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "manual",
         "assisted"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Mode"
     },
     "models": {
      "anyOf": [
       {
        "additionalProperties": {
         "additionalProperties": {
          "type": "string"
         },
         "type": "object"
        },
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Models"
     },
     "prompt_versions": {
      "anyOf": [
       {
        "additionalProperties": {
         "type": "integer"
        },
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Prompt Versions"
     },
     "created_by": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Created By"
     }
    },
    "type": "object",
    "title": "GenerateBody"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "SiteUpdate": {
    "properties": {
     "name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "timezone": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Timezone"
     },
     "wp_url": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 2083,
        "minLength": 1,
        "format": "uri"
       },
       {
        "type": "null"
       }
      ],
      "title": "Wp Url"
     },
     "language": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Language"
     },
     "country": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Country"
     },
     "business_type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Business Type"
     },
     "gsc_property": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Gsc Property"
     },
     "ga4_property": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ga4 Property"
     },
     "mode": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "manual",
         "assisted",
         "autopilot"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Mode"
     }
    },
    "type": "object",
    "title": "SiteUpdate"
   },
...
  }
 }
}